refactor: move to_smallest_unit/from_smallest_unit to helpers.py
The two chain-aware unit converters lived in swap.py as module-level
privates while helpers.py already hosted the sibling to_rao/from_rao.
quote.py also reimplemented to_smallest_unit inline with a raw Decimal
expression.

Move to_smallest_unit/from_smallest_unit into helpers.py next to
to_rao/from_rao, update swap.py to import from helpers, and replace
the inline expression in quote.py with a call to to_smallest_unit.
Drop the now-unused Decimal import from quote.py.

Fixes #81
