+++ allways/cli/swap_commands/helpers.py
@@ -171,6 +171,27 @@ def from_rao(amount_rao: int) -> float:
     return amount_rao / TAO_TO_RAO
 
 
+def to_smallest_unit(amount: float, chain_id: str) -> int:
+    """Convert a human-readable amount to the smallest unit for a chain.
+
+    Uses Decimal to avoid IEEE 754 float artifacts (e.g. 0.1 * 10^9 = 99999999).
+    """
+    from decimal import Decimal
+
+    from allways.chain_defs import get_chain
+
+    chain = get_chain(chain_id)
+    return int(Decimal(str(amount)) * (10**chain.decimals))
+
+
+def from_smallest_unit(amount: int, chain_id: str) -> float:
+    """Convert from smallest unit to human-readable amount."""
+    from allways.chain_defs import get_chain
+
+    chain = get_chain(chain_id)
+    return amount / (10**chain.decimals)
+
+
 def load_cli_config() -> dict:
     """Load CLI configuration from ~/.allways/config.json."""
     if not CONFIG_FILE.exists():
